+++ clothing/src/main/java/m2m_phase2/clothing/clothing/data/model/PaymentM.java
@@ -16,7 +16,7 @@ public class PaymentM {
     private int sysPaymentId;
     private String paymentId;
     private String payerId;
-    private double totalAmount;
+    private Float totalAmount;
     private String currency;
     private String method;
     private String intent;
